+++ fromDistinctAscArray.d.ts
@@ -1,0 +1,5 @@
+import type {ImmutableTree} from './types';
+
+export default function fromDistinctAscArray<T>(
+  array: ReadonlyArray<T>,
+): ImmutableTree<T> | null;