+++ index.d.ts
@@ -3,6 +3,7 @@ export {default as equals} from './equals';
 export {default as find} from './find';
 export {default as findNext} from './findNext';
 export {default as findPrev} from './findPrev';
+export {default as fromDistinctAscArray} from './fromDistinctAscArray';
 export {
   NOOP,
   REPLACE,