+++ index.mjs
@@ -5,6 +5,7 @@ export {default as equals} from './equals.mjs';
 export {default as find} from './find.mjs';
 export {default as findNext} from './findNext.mjs';
 export {default as findPrev} from './findPrev.mjs';
+export {default as fromDistinctAscArray} from './fromDistinctAscArray.mjs';
 export {
   NOOP,
   REPLACE,